requirements.txt
anthropic>=0.40.0
psycopg2-binary>=2.9.0
apscheduler>=3.10.0
websockets>=12.0
python-dotenv>=1.0.0
